Plots for release
- separated plot.py file
- plotting indicator options in config
