+++ config.json
@@ -21,5 +21,12 @@
   "fee": 0.25,
   "strategy-name": "MyStrategy",
   "strategies-folder": "strategies",
-  "plots": "True"
+  "plots": "True",
+  "plot_indicators1": [
+    "ema5",
+    "ema21"
+  ],
+  "plot_indicators2": [
+    "rsi"
+  ]
 }
